+++ NEWS.md
@@ -2,6 +2,7 @@
 * Plot in djpr_plot_server() does not resize when browser height changes
 * Added check_box_selected arg to djpr_plot_server()
 * Render download button on UI side
+* Use body(plot_function) as a cache key
 
 # djprshiny 0.0.5
 * New `interactive` argument to djpr_plot_server()